requirements.txt
Flask
python-telegram-bot
openai
python-dotenv